backend/requirements.txt
asana
cohere
fastapi
openai
python-dotenv
pydantic
uvicorn
